Dancer Procs Module
