+++ locale/de/messages.json
@@ -188,6 +188,10 @@
   "core.weaving.title": "",
   "core.weaving.why": "{0, plural, =1 {# Beispiel für falsches Weaving} other {# Beispiele für falsches Weaving}}",
   "dnc.about.description": "",
+  "dnc.procs.suggestions.drops.content": "",
+  "dnc.procs.suggestions.drops.why": "",
+  "dnc.procs.suggestions.overwrite.content": "",
+  "dnc.procs.suggestions.overwrite.why": "",
   "drg.about.description": "",
   "drg.about.description.warning": "",
   "drg.blood.checklist.description": "",
