+++ locale/en/messages.json
@@ -188,6 +188,10 @@
   "core.weaving.title": "Weaving Issues",
   "core.weaving.why": "{0, plural, =1 {# instance of incorrect weaving} other {# instances of incorrect weaving}}",
   "dnc.about.description": "\nYou can write the description for the job here. It supports a superset of\n[markdown](https://commonmark.org/).\n",
+  "dnc.procs.suggestions.drops.content": "Avoid dropping your procs unless absolutely necessary, as doing so leads to a dps loss unless under very specific circumstances.",
+  "dnc.procs.suggestions.drops.why": "You dropped {droppedProcs, plural, one {# proc} other {# procs}}",
+  "dnc.procs.suggestions.overwrite.content": "Avoid overwriting your procs, as doing so is a major dps loss.",
+  "dnc.procs.suggestions.overwrite.why": "You overwrote {0, plural, one {# proc} other {# procs}}",
   "drg.about.description": "<0>This analyzer aims to help you beat the sterotypes, stay off the floor, and dish out some big juicy numbers. As the DRG rotation is pretty strict, the suggestions you see will focus mostly on keeping that rotation flowing smoothly, as well as picking out issues related to your Life of the Dragon windows.</0>",
   "drg.about.description.warning": "<0>The module is a work in progress.</0> It currently covers very few aspects of DRG gameplay, and while the analysis below should be reasonably accurate, this system <1>is</1> still in development, and may get a little mixed up sometimes. If you notice any issues, or have any concerns, please drop by our Discord channel or report a bug on our github repository!",
   "drg.blood.checklist.description": "<0/> is at the heart of the DRG rotation and should be kept up at all times. Without it, your jumps are weakened and you can't use <1/>.",
