+++ locale/fr/messages.json
@@ -188,6 +188,10 @@
   "core.weaving.title": "",
   "core.weaving.why": "{0, plural, =1 {# exemple de mauvais weaving} other {# exemples de mauvais weaving}}",
   "dnc.about.description": "",
+  "dnc.procs.suggestions.drops.content": "",
+  "dnc.procs.suggestions.drops.why": "",
+  "dnc.procs.suggestions.overwrite.content": "",
+  "dnc.procs.suggestions.overwrite.why": "",
   "drg.about.description": "<0>Ce module a pour but de vous aider à contrer les stéréotypes, rester loin du plancher, et sortir de beaux gros nombres. Étant donné que le Chevalier Dragon possède un rotation très stricte, les suggestions que vous verrez seront présentées de manière à ce que cette rotation demeure fluide, et aussi vous montrer les problèmes existant en lien avec Vie du Dragon.</0>",
   "drg.about.description.warning": "<0>Ce module est en cours de développement.</0> Il couvre présentement peu d'aspects de la manière de jouer du Chevalier Dragon, et bien que les spécifications ci-dessous devraient être raisonnablement précises, ce système <1>est</1> encore en cours de développement et peut certaines fois être un peu sens dessus dessous. Si vous voyez un problème ou avez des questions/requêtes spécifiques, veuillez nous contacter via notre Discord channel ou signalez-nous un bogue sur notre dépôt github!",
   "drg.blood.checklist.description": "<0/> est au cœur de la rotation du DRG et doit être maintenu à tout moment. Sans ça, vos Sauts sont affaiblis et vous ne pouvez plus utiliser <1/>.",
