+++ locale/ja/messages.json
@@ -188,6 +188,10 @@
   "core.weaving.title": "挟み込みの問題",
   "core.weaving.why": "{0, plural, other {不適切な挟み込み回数：#回}}",
   "dnc.about.description": "",
+  "dnc.procs.suggestions.drops.content": "",
+  "dnc.procs.suggestions.drops.why": "",
+  "dnc.procs.suggestions.overwrite.content": "",
+  "dnc.procs.suggestions.overwrite.why": "",
   "drg.about.description": "この分析は、先入観から脱却することを手伝うこと、床から離れるのを助け、大きなジューシーな数字を皿洗いすることを目指しています。 竜騎士のスキル回しはきっちり決まっているため、ここでの助言は、その回しをスムーズに保つこと、そしてあなたの竜の蒼血に関連する問題を解決することに主に焦点を当てます。",
   "drg.about.description.warning": "<0>モジュールは進行中の作業です。</0>現在のところ、竜騎士のゲームプレイのごくわずかな側面しかカバーしていませんが、\n以下の分析は合理的に正確であるべきですが、このシステム<1>はまだ開発中ですが、ときどき混乱することがあります。問題が発生した場合、または懸念がある場合は、Discordチャンネルにお立ち寄りいただくか、githubリポジトリにバグを報告してください。",
   "drg.blood.checklist.description": "<0/>は竜騎士にとってスキル回しの中心であり、切らさず常に維持してスキルを回すことが大事です。これを無視するとジャンプとスパインダイブの威力が低く、<1/>の使用も不可能となります。",
