+++ locale/ko/messages.json
@@ -188,6 +188,10 @@
   "core.weaving.title": "논글쿨기 초과 문제",
   "core.weaving.why": "{0, plural, other {부적절하게 끼워넣은 스킬횟수：#회}}",
   "dnc.about.description": "",
+  "dnc.procs.suggestions.drops.content": "",
+  "dnc.procs.suggestions.drops.why": "",
+  "dnc.procs.suggestions.overwrite.content": "",
+  "dnc.procs.suggestions.overwrite.why": "",
   "drg.about.description": "<0>이 분석기는 여러분이 용기사에 대한 편견을 이겨내고, 바닥에 눕지 않으면서 미터기의 숫자를 높게 올리도록 돕는 것을 목표로 하고 있습니다. 용기사의 딜사이클은 상당히 엄격하기 때문에, 여러분이 보는 제안들은 주로 붉은 용혈과 관련된 문제들을 골라내는 것뿐만 아니라, 그 딜사이클이 원활하게 돌아갈 수 있는 것에 초점이 맞춰져있습니다.</0>",
   "drg.about.description.warning": "<0>이 모듈은 현재 작업중입니다.</0> 현재 용기사 운용의 극히 일부만을 다루고 있으며, 어느 정도 정확한 분석을 제공하고 있지만 시스템이 아직 개발중이기 때문에 조금 혼란스러우실수 있습니다. 만약 문제가 있거나 우려되는 부분이 있다면 저희 디스코드 채널로 오시거나 github에 버그를 신고해주시길 바랍니다.",
   "drg.blood.checklist.description": "<0/>은 용기사에게 있어서 딜사이클의 중심이며, 끊임없이 유지해 스킬을 돌리는 것이 중요합니다. 용혈이 끊기면 점프의 위력이 낮아지며 <1/>의 사용도 불가능해집니다.",
