+++ locale/zh/messages.json
@@ -188,6 +188,10 @@
   "core.weaving.title": "能力技穿插问题",
   "core.weaving.why": "{0, plural, other {# 次错误插入}}",
   "dnc.about.description": "",
+  "dnc.procs.suggestions.drops.content": "",
+  "dnc.procs.suggestions.drops.why": "",
+  "dnc.procs.suggestions.overwrite.content": "",
+  "dnc.procs.suggestions.overwrite.why": "",
   "drg.about.description": "<0>此分析工具的目的是帮助你克服刻板印象，告别躺尸并打出赏心悦目的数据。因为龙骑士的循环非常严格，这里的建议主要着眼于如何保持流畅的循环，以及找出关于你的红莲龙血的问题。</0>",
   "drg.about.description.warning": "<0>本模块还在开发中。</0>目前只涉及了龙骑士手法的一小部分，虽然下面的分析应是相当准确的，但这个系统<1>仍在</1>在开发中，也许有时会出一些问题。如果你需要反映问题或相关事项，请加入我们的Discord频道或者在GitHub仓库报告Bug。",
   "drg.blood.checklist.description": "<0/>是龙骑士循环的核心，应该始终维持。没有它时，你的跳跃威力会减弱，且无法使用<1/>。",
